+++ app/controllers/integral/posts_controller.rb
@@ -1,7 +1,7 @@
 module Integral
   # Posts controller
   class PostsController < BlogController
-    prepend_before_action :find_post, only: [:show]
+    before_action :find_post, only: [:show]
     before_action :find_related_posts, only: [:show]
     after_action :increment_post_count, only: [:show]
 
@@ -9,8 +9,9 @@ module Integral
     # List blog posts
     def index
       add_breadcrumb I18n.t('integral.breadcrumbs.blog'), nil
-      @latest_post = Integral::Post.published.order('published_at DESC').first&.decorate
-      @posts = Integral::Post.published.includes(:image, :user).order('published_at DESC').paginate(page: params[:page])
+
+      @latest_post = Integral::Post.published.where(locale: I18n.locale).order('published_at DESC').first&.decorate
+      @posts = Integral::Post.published.where(locale: I18n.locale).includes(:image, :user).order('published_at DESC').paginate(page: params[:page])
       @posts = @posts.where.not(id: @latest_post.id) if @latest_post
     end
 
@@ -33,6 +34,23 @@ module Integral
 
     private
 
+    def alternative_urls
+      alternative_urls = {
+        I18n.locale.to_s => canonical_url
+      }
+
+      if action_name == 'show'
+        @post.alternates.published.each do |alternate|
+          alternative_urls[alternate.locale] = alternate.frontend_url
+        end
+      elsif
+        Integral.frontend_locales.reject{ |l| l == I18n.locale}.each do |locale|
+          alternative_urls[locale.to_s] = self.send("posts_#{locale}_url")
+        end
+      end
+      alternative_urls
+    end
+
     def increment_post_count
       @post.increment_count!(request.ip)
     end
@@ -53,9 +71,9 @@ module Integral
 
     def find_post
       @post = if current_user.present?
-                Integral::Post.friendly.find(params[:id]).decorate
+                Integral::Post.where(locale: params[:locale]).friendly.find(params[:id]).decorate
               else
-                Integral::Post.friendly.published.find(params[:id]).decorate
+                Integral::Post.where(locale: params[:locale]).friendly.published.find(params[:id]).decorate
               end
 
       @post.decorate
@@ -63,7 +81,7 @@ module Integral
       # If an old id or a numeric id was used to find the record, then
       # the request path will not match the post_path, and we should do
       # a 301 redirect that uses the current friendly id.
-      redirect_to post_url(@post), status: :moved_permanently if request.path != post_path(@post)
+      redirect_to post_url(@post.slug), status: :moved_permanently if request.path != post_path(@post.slug)
     end
   end
 end